new Document model to make editor cleaner
The editor view shouldn’t need to know how to serialize things, so I’ve
moved that code (almost verbatim) to a new Document model object
